Refactor

+++ db/migrate/20160411180135_create_ticket_types.rb
@@ -1,7 +1,7 @@
 class CreateTicketTypes < ActiveRecord::Migration
   def change
     create_table :ticket_types do |t|
-      t.decimal :cost
+      t.integer :cost
       t.string :level
       t.integer :event_id
       t.timestamps null: false
